requirements.txt
contextily==1.6.2
geopandas==1.0.1
geopy==2.4.1
h3==4.2.2
httpx==0.27.2
langchain_chroma==0.2.0
langchain_community==0.3.14
langchain_core==0.3.29
langchain_huggingface==0.1.2
langchain_openai==0.2.14
langgraph==0.2.61
matplotlib==3.10.1
numpy==1.26.4
openai==1.54.5
pandas==2.2.3
pycountry==24.6.1
pydantic==2.9.2
pyproj==3.7.1
python-dotenv==1.0.1
Requests==2.32.3
Shapely==2.0.7
stanza==1.10.1
tqdm==4.66.6
GitPython==3.1.44
seaborn==0.13.2
scikit-learn==1.6.1
scipy==1.15.2